discord-bot/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=11.0.3
pydub>=0.25.1
ffmpeg-python>=0.2.0
PyNaCl>=1.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1
gTTS>=2.5.0